pyproject.toml
[project]
name = "mudra"
version = "0.1.0"
description = ""
authors = [
    {name = "OMAR",email = "omar@localhost.local"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mediapipe (>=0.10.21,<0.11.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "customtkinter (>=5.2.2,<6.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
